Refactoring to make partitioning code reusable
