+++ storage/tokudb/ha_tokudb.h
@@ -1010,8 +1010,6 @@ private:
         const std::string &partition_info_str);
     MY_NODISCARD int rename_non_partitioned_table(
         const char *from, const char *to);
-    MY_NODISCARD partition_info *parse_partition_info(
-        const std::string &partition_info_str);
 public:
     // mysql sometimes retires a txn before a cursor that references the txn is closed.
     // for example, commit is sometimes called before index_end.  the following methods
@@ -1045,38 +1043,6 @@ private:
 static inline bool key_is_clustering(const KEY *key) {
     return key->flags & HA_CLUSTERING;
 }
-
-template<typename Fn>
-class exit_scope
-{
-public:
-    exit_scope(const Fn& fn):m_fn(fn) {}
-    ~exit_scope(){ m_fn(); }
-    // Make sure NRVO is used (mandatory since C++17)
-    exit_scope(const exit_scope&);
-
-private:
-    exit_scope() = delete;
-    exit_scope& operator =(const exit_scope&) = delete;
-
-    Fn m_fn;
-};
-
-class exit_scope_creator
-{
-public:
-    template<typename Fn>
-    exit_scope<Fn> operator << (const Fn &fn)
-    {
-        return exit_scope<Fn>(fn);
-    }
-};
-
-#define EXIT_SCOPE_CREATE_UNIQ_NAME2(line) exit_scope_guard_##line
-#define EXIT_SCOPE_CREATE_UNIQ_NAME(line) EXIT_SCOPE_CREATE_UNIQ_NAME2(line)
-#define EXIT_SCOPE \
-    const auto  &EXIT_SCOPE_CREATE_UNIQ_NAME(__LINE__) = \
-    exit_scope_creator() << [&]()
 
 #endif // _HA_TOKUDB_H
